Add files via upload
This program will now clean the data and create a .csv file for the data. 

I was getting annoyed with how long this was becoming and so i split up the program. This is typical practice for me, however, at the beginning of every project it has hard to know how long a program will get. Hence why I am splitting up the program now.